etc/ssh.csh: Fix a bug (not csh-specific) in ssh-agent-dup
The bug was that we were depending on $owner to be the OS-specific
stat flags but then we overwrite it with the results of the first-
use. This meant that we only ever detected one agent in ssh-agent-
dup(). By fixing this bug, ssh-agent-dup properly detects multiple
agents forwarded into a box owned by the current user without the
additional use of super-powers and the -a flag.